Tests: break up parser unit tests into modules
